git-svn-id: https://subversion.assembla.com/svn/cmosher@9151 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ epple2qt/src/analogtv.h
@@ -42,7 +42,7 @@ private:
 	int isig;
 	bool noise;
 	DisplayType type;
-	unsigned long long int rrr;
+	unsigned long /* TODO long*/ int rrr;
 
 	static int* rcb;
